requirements.txt
requests
litellm
PyPDF2
PyMuPDF
python-dotenv
torch
transformers
accelerate
sentencepiece
pyarrow
sentence-transformers
scikit-learn
numpy
scipy
matplotlib